Implemented google oauth and changed github workflow

+++ server/src/controllers/user.rs
@@ -2,19 +2,13 @@ use crate::{
     app::AppState,
     auth::AuthenticatedUser,
     controllers::helper::OkResponse,
-    models::user::{NewUser, User},
+    models::user::User,
     util::errors::{AppError, AppResult},
     views::EncodableUser,
 };
 use actix_web::web::{self, Json};
-use argon2::{
-    Argon2,
-    password_hash::{PasswordHasher, SaltString, rand_core::OsRng},
-};
-use base64::{Engine, engine::general_purpose};
 use diesel::{ExpressionMethods, result::Error::NotFound};
 use diesel_async::RunQueryDsl;
-use regex::Regex;
 use serde::{Deserialize, Serialize};
 use utoipa::ToSchema;
 use uuid::Uuid;
@@ -29,6 +23,7 @@ pub struct GetUsersResponse {
     path = "/api/v1/users",
     responses(
         (status = 200, description = "Successful Response", body = GetUsersResponse),
+        (status = 403, description = "Insufficient permissions"),
     ),
     security(
         ("jwt" = [])
@@ -38,6 +33,10 @@ pub async fn get_users(
     authenticated: AuthenticatedUser,
     state: web::Data<AppState>,
 ) -> AppResult<Json<GetUsersResponse>> {
+    if !authenticated.is_admin() {
+        return Err(AppError::Forbidden);
+    }
+
     let mut conn = state.db_connection().await?;
 
     let result = User::get_all(&mut conn).await;
@@ -60,69 +59,6 @@ pub async fn get_users(
     }
 }
 
-#[derive(Deserialize, Serialize, ToSchema)]
-pub struct CreateUserBody {
-    pub username: String,
-    pub email: String,
-    pub password: String,
-}
-
-fn is_valid_email(email: &str) -> bool {
-    let re = Regex::new(r"^[a-zA-Z0-9.!#$%&'*+/=?^_`{|}~-]+@[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?(?:\.[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?)*$").unwrap();
-    re.is_match(email)
-}
-
-#[utoipa::path(
-    tag = "users",
-    post,
-    path = "/api/v1/users",
-    responses(
-        (status = 200, description = "Successful Response", body = OkResponse),
-    )
-)]
-pub async fn create_user(
-    new_user_data: web::Json<CreateUserBody>,
-    state: web::Data<AppState>,
-) -> AppResult<OkResponse> {
-    let new_user_data = new_user_data.into_inner();
-
-    let salt = SaltString::generate(&mut OsRng);
-    let salt_bytes: Vec<u8> = general_purpose::STANDARD_NO_PAD
-        .decode(salt.as_str())
-        .unwrap();
-
-    let argon2 = Argon2::default();
-
-    let password_hash: String;
-    if let Ok(hash) = argon2.hash_password(new_user_data.password.as_bytes(), &salt) {
-        password_hash = hash.to_string();
-    } else {
-        return Err(AppError::InternalError);
-    }
-
-    // check email validity
-    if !is_valid_email(&new_user_data.email) {
-        return Err(AppError::BadRequest("Malformed email address".to_string()));
-    }
-
-    let new_user = NewUser {
-        username: Some(&new_user_data.username),
-        email: Some(&new_user_data.email),
-        password_hash: Some(&password_hash),
-        password_salt: Some(&salt_bytes),
-        avatar: None,
-    };
-
-    let mut conn = state.db_connection().await?;
-
-    let result = new_user.insert(&mut conn).await;
-
-    match result {
-        Ok(_) => Ok(OkResponse::new()),
-        Err(_) => Err(AppError::InternalError),
-    }
-}
-
 #[derive(Serialize, Deserialize, ToSchema)]
 pub struct GetUserResponse {
     pub user: EncodableUser,
@@ -134,7 +70,9 @@ pub struct GetUserResponse {
     path = "/api/v1/users/{user_id}",
     responses(
         (status = 200, description = "Successful Response", body = GetUserResponse),
+        (status = 403, description = "Insufficient permissions"),
         (status = 404, description = "User Not Found"),
+        (status = 500, description = "Internal server error"),
     ),
     security(
         ("jwt" = [])
@@ -145,6 +83,10 @@ pub async fn get_user(
     path: web::Path<Uuid>,
     state: web::Data<AppState>,
 ) -> AppResult<Json<GetUserResponse>> {
+    if !authenticated.is_admin() {
+        return Err(AppError::Forbidden);
+    }
+
     let user_id = path.into_inner();
 
     let mut conn = state.db_connection().await?;
@@ -171,6 +113,9 @@ pub async fn get_user(
     path = "/api/v1/users/{user_id}",
     responses(
         (status = 200, description = "Successful Response", body = OkResponse),
+        (status = 403, description = "Insufficient permissions"),
+        (status = 404, description = "User not found"),
+        (status = 500, description = "Internal server error"),
     ),
     security(
         ("jwt" = [])
@@ -183,13 +128,17 @@ pub async fn delete_user(
 ) -> AppResult<OkResponse> {
     let user_id = path.into_inner();
 
+    if !authenticated.is_admin() && authenticated.user_id != user_id {
+        return Err(AppError::Forbidden);
+    }
+
     let mut conn = state.db_connection().await?;
 
     let result = User::delete(&mut conn, &user_id).await;
 
     match result {
         Ok(_) => Ok(OkResponse::new()),
-        Err(NotFound) => Err(AppError::BadRequest("User not found".to_string())),
+        Err(NotFound) => Err(AppError::NotFound),
         Err(_) => Err(AppError::InternalError),
     }
 }
@@ -208,6 +157,9 @@ pub struct UpdateInformationBody {
     path = "/api/v1/users/{user_id}",
     responses(
         (status = 200, description = "Successful Response", body = OkResponse),
+        (status = 403, description = "Insufficient permissions"),
+        (status = 404, description = "User not found"),
+        (status = 500, description = "Internal server error"),
     ),
     security(
         ("jwt" = [])
@@ -221,6 +173,10 @@ pub async fn update_user(
 ) -> AppResult<OkResponse> {
     let user_id = path.into_inner();
 
+    if !authenticated.is_admin() && authenticated.user_id != user_id {
+        return Err(AppError::Forbidden);
+    }
+
     let mut conn = state.db_connection().await?;
 
     use crate::schema::users::dsl::*;
@@ -233,7 +189,7 @@ pub async fn update_user(
             .await;
 
         if result == Err(NotFound) {
-            return Err(AppError::BadRequest("User not found".to_string()));
+            return Err(AppError::NotFound);
         } else if result.is_err() {
             return Err(AppError::InternalError);
         }
@@ -247,7 +203,7 @@ pub async fn update_user(
             .await;
 
         if result == Err(NotFound) {
-            return Err(AppError::BadRequest("User not found".to_string()));
+            return Err(AppError::NotFound);
         } else if result.is_err() {
             return Err(AppError::InternalError);
         }
